Refactor native calls out of PrefetchLibGit2Repo into platform specific functions.
