--- GVFS/FastFetch/packages.config
@@ -1,6 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="CommandLineParser" version="2.1.1-beta" targetFramework="net461" />
-  <package id="LibGit2Sharp.NativeBinaries" version="1.0.165" targetFramework="net461" />
-  <package id="StyleCop.Analyzers" version="1.0.2" targetFramework="net461" developmentDependency="true" />
-</packages>
