+++ GVFS/GVFS.Build/GVFS.cs.props
@@ -14,6 +14,11 @@
     <CodeAnalysisRuleSet>$(MSBuildThisFileDirectory)\GVFS.ruleset</CodeAnalysisRuleSet>
   </PropertyGroup>
 
+  <PropertyGroup Condition="'$(MSBuildRuntimeType)' == 'Core'">
+    <IsOSX Condition="'$([System.Runtime.InteropServices.RuntimeInformation]::IsOSPlatform($([System.Runtime.InteropServices.OSPlatform]::OSX)))' == 'true'">true</IsOSX>
+    <IsLinux Condition="'$([System.Runtime.InteropServices.RuntimeInformation]::IsOSPlatform($([System.Runtime.InteropServices.OSPlatform]::Linux)))' == 'true'">true</IsLinux>
+  </PropertyGroup>
+  
   <PropertyGroup>
     <DefaultItemExcludes>$(DefaultItemExcludes);StyleCop.Cache;TestResults.xml</DefaultItemExcludes>
   </PropertyGroup>
