+++ GVFS/GVFS.Common/FileSystem/IPlatformFileSystem.cs
@@ -1,4 +1,6 @@
-﻿namespace GVFS.Common.FileSystem
+﻿using GVFS.Common.Tracing;
+
+namespace GVFS.Common.FileSystem
 {
     public interface IPlatformFileSystem
     {
@@ -9,8 +11,8 @@
         bool TryGetNormalizedPath(string path, out string normalizedPath, out string errorMessage);
         void ChangeMode(string path, int mode);
         bool HydrateFile(string fileName, byte[] buffer);
-
         bool IsExecutable(string filePath);
         bool IsSocket(string filePath);
+        unsafe void WriteFile(ITracer tracer, byte* originalData, long originalSize, string destination);
     }
 }
