+++ GVFS/GVFS.Platform.Mac/MacFileSystem.cs
@@ -1,6 +1,8 @@
 ﻿using GVFS.Common;
 using GVFS.Common.FileSystem;
+using GVFS.Common.Tracing;
 using System;
+using System.ComponentModel;
 using System.IO;
 using System.Runtime.InteropServices;
 
@@ -56,11 +58,37 @@ namespace GVFS.Platform.Mac
             return NativeStat.IsSock(statBuffer.Mode);
         }
 
+        public unsafe void WriteFile(ITracer tracer, byte* originalData, long originalSize, string destination)
+        {
+            int fileDescriptor = 1;
+            try
+            {
+                fileDescriptor = NativeFileReader.Open(destination, NativeFileReader.WriteOnly | NativeFileReader.Create);
+                IntPtr result = Write(fileDescriptor, originalData, (IntPtr)originalSize);
+                if (result.ToInt32() == -1)
+                {
+                    throw new Win32Exception(Marshal.GetLastWin32Error());
+                }
+            }
+            catch
+            {
+                tracer.RelatedError("Error writing file {0}. ERRNO: {1}", destination, Marshal.GetLastWin32Error());
+                throw;
+            }
+            finally
+            {
+                NativeFileReader.Close(fileDescriptor);
+            }
+        }
+
         [DllImport("libc", EntryPoint = "chmod", SetLastError = true)]
         private static extern int Chmod(string pathname, int mode);
 
         [DllImport("libc", EntryPoint = "rename", SetLastError = true)]
         private static extern int Rename(string oldPath, string newPath);
+
+        [DllImport("libc", EntryPoint = "write", SetLastError = true)]
+        private static unsafe extern IntPtr Write(int fileDescriptor, void* buf, IntPtr count);
 
         private NativeStat.StatBuffer StatFile(string fileName)
         {
@@ -141,7 +169,10 @@ namespace GVFS.Platform.Mac
 
         private class NativeFileReader
         {
-            private const int ReadOnly = 0x0000;
+            public const int ReadOnly = 0x0000;
+            public const int WriteOnly = 0x0001;
+
+            public const int Create = 0x0200;
 
             public static bool TryReadFirstByteOfFile(string fileName, byte[] buffer)
             {
@@ -160,6 +191,15 @@ namespace GVFS.Platform.Mac
                 return readStatus;
             }
 
+            [DllImport("libc", EntryPoint = "open", SetLastError = true)]
+            public static extern int Open(string path, int flag);
+
+            [DllImport("libc", EntryPoint = "close", SetLastError = true)]
+            public static extern int Close(int fd);
+
+            [DllImport("libc", EntryPoint = "read", SetLastError = true)]
+            public static extern int Read(int fd, [Out] byte[] buf, int count);
+
             private static bool TryReadOneByte(int fileDescriptor, byte[] buffer)
             {
                 int numBytes = Read(fileDescriptor, buffer, 1);
@@ -171,15 +211,6 @@ namespace GVFS.Platform.Mac
 
                 return true;
             }
-
-            [DllImport("libc", EntryPoint = "open", SetLastError = true)]
-            private static extern int Open(string path, int flag);
-
-            [DllImport("libc", EntryPoint = "close", SetLastError = true)]
-            private static extern int Close(int fd);
-
-            [DllImport("libc", EntryPoint = "read", SetLastError = true)]
-            private static extern int Read(int fd, [Out] byte[] buf, int count);
         }
     }
 }
