+++ GVFS/GVFS.Platform.Windows/GVFS.Platform.Windows.csproj
@@ -12,6 +12,7 @@
     <AssemblyName>GVFS.Platform.Windows</AssemblyName>
     <TargetFrameworkVersion>v4.6.1</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
+    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
     <NuGetPackageImportStamp>
     </NuGetPackageImportStamp>
   </PropertyGroup>
